rename: lowercase directory name
